rf: move clear_logs and restart_wazuh_function
